api error fix for github fetch
